nydus-image: avoid duplicated code in builder
Abstract common code in builder as helper functions to reduce duplicated
code.
